refactor: rename bounded contexts to 4-phase lifecycle (#49)
Rename context directories to match the redesigned bounded contexts:
- scheduling -> preparation (merged with existing empty preparation)
- recording -> record
- publishing -> followup
- Add new session context skeleton

Update all import paths accordingly. All 110 unit tests pass.
