+++ backend/contexts/preparation/domain/value_objects.py
