+++ backend/contexts/record/domain/exceptions.py
